+++ frontend/src/locales/en.json
@@ -31,7 +31,8 @@
     "javascript": "// Write your code in JavaScript\n",
     "php": "// Write your code in PHP\n",
     "python": "# Write your code in Pyhton\n",
-    "html": "<!--Write your code in HTML-->\n"
+    "html": "<!--Write your code in HTML-->\n",
+    "java": "// Write your code in JAVA\n"
   },
   "editor": {
     "authAlert": "To edit snippet, you must be its author or copy it",
@@ -343,7 +344,7 @@
     },
     "saveCode": {
       "success": "Snippet successfully saved!",
-      "error" : "Snippet was not saved"
+      "error": "Snippet was not saved"
     }
   }
 }
