+++ frontend/src/locales/ru.json
@@ -31,7 +31,8 @@
     "javascript": "// Здесь будет код на JavaScript\n",
     "php": "// Здесь будет код на PHP\n",
     "python": "# Здесь будет код на Pyhton\n",
-    "html": "<!--Здесь будет код на HTML-->\n"
+    "html": "<!--Здесь будет код на HTML-->\n",
+    "java": "// Здесь будет код на JAVA\n"
   },
   "editor": {
     "authAlert": "Чтобы изменить сниппет, необходимо быть его автором или скопировать его",
@@ -475,7 +476,7 @@
     },
     "saveCode": {
       "success": "Сниппет успешно сохранен!",
-      "error" : "Не удалось сохранить сниппет"
+      "error": "Не удалось сохранить сниппет"
     }
   }
 }
